Move Axis-Formatters to a new util file so we can exclude from coverage the Charts file which is for UI purposes
# Conflicts:
#	app/src/main/java/com/weatherxm/util/Charts.kt
